requirements.txt
tensorflow==2.20.0
keras==3.12.0
numpy==1.26.4
pillow==11.3.0
matplotlib==3.10.8
requests==2.32.3
protobuf==5.28.3